+++ CHANGELOG.md
@@ -5,6 +5,11 @@
   ## **WORK IN PROGRESS**
   * (theimo1221) Update packages
 -->
+
+## **WORK IN PROGRESS**
+
+* (theimo1221) Debounce Zigbee Actuator Persistence
+
 ## 3.0.0-alpha.29 (2024-04-01)
 
 * (theimo1221) Correct isAutomatic detection